backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
openai
python-multipart
websockets
google-generativeai
requests
groq
tavily-python
flask